tag onlpv1-r5.0.12 from ufi-onlpv1-dev branch -------------------------------------------- 01. [All][Common]      - [x] update i40e driver from 2.16.11 to 2.22.8 to support for higher kernels. 02. [S6301-56ST][r1.0.1]      - [x] update ID led color from GREEN to BLUE      - [x] add onlp_set capability for SYS LED 03. [S8901-54XC][r1.0.3]      - [x] refine bios version command      - [x] update tech support script for beta board      - [x] add Delta psu support      - [x] add platform support for higher kernels 04. [S9110-32X][r1.0.2]      - [x] update bsp version from 1.0.1 to 1.0.2 05. [S9300-32D][r1.0.14]      - [x] add execute permission to tech support script 06. [S9301-32D][r1.0.2] ------- No Update 07. [S9301-32DB][r1.0.8]      - [x] add execute permission to tech support script 08. [S9310-32D][r1.0.9] ------- No Update 09. [S9500-22XST][r0.6] ------- No Update 10. [S9500-54CF][r1.0.3]      - [x] correct SFP_TX_FAULT cpld register      - [x] correct SFP_RX_LOS cpld register 11. [S9501-16SMT][r1.0.9] ----- No Update 12. [S9501-18SMT][r1.0.12] ---- No Update 13. [S9501-28SMT][r1.0.14] ---- No Update 14. [S9502-12SM][r1.0.5] ------ No Update 15. [S9502-16SMT][r1.0.5] ----- No Update 16. [S9510-28DC][r1.0.17] ----- No Update 17. [S9510-30XC][r1.0.5] ------ No Update 18. [S9600-28DX][r1.0.0] ------ No Update 19. [S9600-30DX][r1.0.6]      - [x] refine bios version command      - [x] add BIT support      - [x] remove unused variable CPLD_TEST in cpld driver      - [x] add platform support for higher kernels 20. [S9600-32X][r1.0.15]      - [x] add mu version info      - [x] add platform support for higher kernels 21. [S9600-48X][r1.0.6] ------- No Update 22. [S9600-56DX][r1.0.2] ------ No Update 23. [S9600-64X][r1.0.16]      - [x] fixed wrong logging path in tech support script      - [x] add platform support for higher kernels 24. [S9600-72XC][r1.0.18]      - [x] add mu version info      - [x] add platform support for higher kernels 25. [S9610-36D][r1.0.2]      - [x] refine bios version command 26. [S9700-23D][r4.2.24]      - [x] add mu version info      - [x] refine bios version command      - [x] add platform support for higher kernels 27. [S9700-53DX][r4.2.24]      - [x] add mu version info      - [x] refine bios version command      - [x] add platform support for higher kernels 28. [S9701-78DC][r1.0.11] ----- No Update 29. [S9701-82DC][r1.0.11]      - [x] add mu version info      - [x] add platform support for higher kernels 30. [S9705-48D][r4.2.24]      - [x] add mu version info      - [x] refine bios version command      - [x] update board info description and cpu thermal sensor number in tech support script      - [x] add platform support for higher kernels 31. [S9710-76D][r1.0.20]      - [x] add mu version info      - [x] refine bios version command      - [x] add platform support for higher kernels 33. [S9100-32X][r4.2.15] ------ No Update 33. [S9180-32X][r4.2.20]      - [x] add fan status attribute (direction) 34. [S9230-64X][r4.2.15] ------ No Update 35. [S9280-64X][r4.2.18] ------ No Update 36. [S9280-64X-4BWB][r4.2.16] - No Update

+++ packages/platforms/ufispace/x86-64/modules/builds/src/i40e/i40e_dcb.c
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: GPL-2.0
-/* Copyright(c) 2013 - 2021 Intel Corporation. */
+/* Copyright(c) 2013 - 2022 Intel Corporation. */
 
 #include "i40e_adminq.h"
 #include "i40e_prototype.h"
@@ -905,7 +905,7 @@ i40e_status i40e_init_dcb(struct i40e_hw *hw, bool enable_mib_change)
 		return I40E_ERR_NOT_READY;
 
 	/* Get the LLDP AdminStatus for the current port */
-	adminstatus = lldp_cfg.adminstatus >> (hw->port * 4);
+	adminstatus = (u8)(lldp_cfg.adminstatus >> (hw->port * 4));
 	adminstatus &= 0xF;
 
 	/* LLDP agent disabled */
